requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
Bottleneck==1.3.7
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.4
cycler==0.11.0
decorator==5.1.1
distlib==0.3.7
executing==1.2.0
filelock==3.12.2
fonttools==4.42.0
greenlet==2.0.2
icecream==2.1.3
idna==3.4
iniconfig==2.0.0
ipython==8.15.0
jedi==0.19.0
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numexpr==2.8.7
numpy==1.26.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
playwright==1.36.0
plotly==5.15.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.8
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyee==9.0.4
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.6.2
tenacity==8.2.2
traitlets==5.10.1
typing_extensions==4.8.0
urllib3==2.0.5
virtualenv==20.24.3
wcwidth==0.2.6
